added some tests, refactored examples

+++ simpleGauss/main_simple_gauss_measure_cpu.cpp
@@ -14,12 +14,12 @@ int main(int argc, char *argv[])
     }
 
     size_t n = strtol(argv[1], nullptr, 0);
-    float low = -100;
-    float high = 100;
+    double low = -100;
+    double high = 100;
 
-    SimpleGaussCPU<DenseMat<float>, float> cpuSolver;
+    SimpleGaussCPU<DenseMat<double>, double> cpuSolver;
 
-    DenseMat<float> mat(DenseMat<float>::genRandMat(n, n + 1, low, high));
+    DenseMat<double> mat(DenseMat<double>::genRandMat(n, n + 1, low, high));
     auto cpuSolution(cpuSolver.solve(mat));
     return 0;
 }
