+++ simpleGauss/main_simple_gauss_measure_cuda.cpp
@@ -13,12 +13,12 @@ int main(int argc, char *argv[])
         return 0;
     }
     size_t n = strtol(argv[1], nullptr, 0);
-    float low = -100;
-    float high = 100;
+    double low = -100;
+    double high = 100;
 
-    SimpleGaussCuda<float> cudaSolver;
+    SimpleGaussCuda<double> cudaSolver;
 
-    DenseMat<float> mat(DenseMat<float>::genRandMat(n, n + 1, low, high));
+    DenseMat<double> mat(DenseMat<double>::genRandMat(n, n + 1, low, high));
     auto cudaSolution(cudaSolver.solve(mat));
     return 0;
 }
